DUNE/Utils/Codecs: Fixed some compilation warnings.

+++ src/DUNE/Utils/Codecs/CodedEstimatedState.hpp
@@ -85,7 +85,7 @@ namespace DUNE
 
           uint8_t* ptr = (uint8_t*)&frame->data[1];
 
-          uint16_t length = frame->data.size() - 1;
+          uint16_t length = (uint16_t)frame->data.size() - 1;
           ptr += IMC::deserialize(coded.lat, ptr, length);
           ptr += IMC::deserialize(coded.lon, ptr, length);
           ptr += IMC::deserialize(coded.alt, ptr, length);
